[BPK-4276] Restructure native pages
