+++ docs/src/pages/NavigationBarPage/NativeNavigationBarPage.js
