+++ docs/src/pages/NudgerPage/NudgerPage.js
@@ -23,7 +23,8 @@ import React from 'react';
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
 import Web from '../WebNudgersPage';
-import Native from '../NativeNudgerPage';
+
+import Native from './NativeNudgerPage';
 
 const Page = () => (
   <DocsPageWrapper
