+++ docs/src/pages/ProgressPage/NativeProgressPage.js
