+++ docs/src/pages/RatingPage/NativeRatingPage.js
