+++ docs/src/pages/RatingPage/RatingPage.js
@@ -24,8 +24,9 @@ import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
 import AndroidRating from '../AndroidRatingPage';
 import IOSRating from '../IOSRatingPage';
-import NativeRating from '../NativeRatingPage';
 import WebRating from '../WebRatingPage';
+
+import NativeRating from './NativeRatingPage';
 
 const Page = () => (
   <DocsPageWrapper
