+++ docs/src/pages/SectionListPage/SectionListPage.js
@@ -23,7 +23,8 @@ import React from 'react';
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
 import Web from '../WebSectionListPage';
-import Native from '../NativeSectionListPage';
+
+import Native from './NativeSectionListPage';
 
 const Page = () => (
   <DocsPageWrapper
