+++ docs/src/pages/SelectPage/SelectPage.js
@@ -22,8 +22,9 @@ import React from 'react';
 
 import DocsPageWrapper from '../../components/DocsPageWrapper';
 import IntroBlurb from '../../components/IntroBlurb';
-import NativeSelects from '../NativeSelectPage';
 import WebSelects from '../WebSelectPage';
+
+import NativeSelects from './NativeSelectPage';
 
 const blurb = [
   <IntroBlurb>
