+++ docs/src/pages/SpinnerPage/NativeSpinnerPage.js
