+++ docs/src/pages/SpinnerPage/SpinnerPage.js
@@ -25,7 +25,8 @@ import IntroBlurb from '../../components/IntroBlurb';
 import Android from '../AndroidSpinnerPage';
 import IOS from '../IOSSpinnerPage';
 import Web from '../WebSpinnersPage';
-import Native from '../NativeSpinnerPage';
+
+import Native from './NativeSpinnerPage';
 
 const blurb = [
   <IntroBlurb>
