+++ docs/src/pages/StarRatingPage/NativeStarRatingPage.js
