+++ docs/src/pages/SwitchPage/NativeSwitchPage.js
